pyproject.toml
[project]
name = "restaurant-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.9.1",
    "black>=25.1.0",
    "Django==5.2.4",
    "django-cors-headers>=4.7.0",
    "django-filter==25.1",
    "django-user-agents>=0.4.0",
    "djangorestframework==3.16.0",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-forms>=1.1.8",
    "drf-spectacular>=0.28.0",
    "filetype>=1.2.0",
    "gunicorn>=23.0.0",
    "nepali-datetime>=1.0.8.4",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple==3.8",
    "requests>=2.32.4",
    "rest-framework-simplejwt>=0.0.2",
    "ruff>=0.12.9",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
    "uvicorn>=0.35.0",
]
